Create Hackathon2021
This is our telegram bot that we developing. We are beginner coder that are still learning how to code on the telegram bot to make ti easy and covenient for people to use. Our bot is not working as of now. However this is a foundation of what we are planning to do in the future on developing easy access on the information on covid 19 and even other information in the future.